+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to
 ### Added
 
 - ✨(backend) add documents/all endpoint with descendants #1553
+- ✅(export) add PDF regression tests #1762
 
 ## [4.3.0] - 2026-01-05